requirements.txt
Flask
twilio
python-dotenv
openai-whisper
faster-whisper
google-cloud-speech
openai